+++ src/storage/store/async/google_cloud_store.rs
@@ -1,0 +1,18 @@
+//! Google Cloud stores.
+
+use crate::{object_store_impl, storage::StorageError};
+
+/// A Google Cloud Storage store.
+#[derive(Debug)]
+pub struct AsyncGoogleCloudStore {
+    object_store: object_store::gcp::GoogleCloudStorage,
+}
+
+impl AsyncGoogleCloudStore {
+    /// Create a new amazon S3 store.
+    pub fn new(object_store: object_store::gcp::GoogleCloudStorage) -> Self {
+        Self { object_store }
+    }
+}
+
+object_store_impl!(AsyncGoogleCloudStore, object_store);